Moves further components to separate packages
