+++ components/src/jsMain/kotlin/dev/fritz2/components/forms/control/component.kt
@@ -10,6 +10,7 @@ import dev.fritz2.components.foundations.Component
 import dev.fritz2.components.foundations.ComponentProperty
 import dev.fritz2.components.foundations.FormMixin
 import dev.fritz2.components.foundations.FormProperties
+import dev.fritz2.components.inputField.InputFieldComponent
 import dev.fritz2.components.slider.SliderComponent
 import dev.fritz2.components.validation.ComponentValidationMessage
 import dev.fritz2.components.validation.Severity
