+++ components/src/jsMain/kotlin/dev/fritz2/components/nav.kt
@@ -3,6 +3,7 @@ package dev.fritz2.components
 import dev.fritz2.components.foundations.Component
 import dev.fritz2.components.foundations.ComponentProperty
 import dev.fritz2.components.foundations.DynamicComponentProperty
+import dev.fritz2.components.icon.IconComponent
 import dev.fritz2.dom.Listener
 import dev.fritz2.dom.html.RenderContext
 import dev.fritz2.styling.*
